requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
openpyxl==3.1.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0 